refactor(runner): split workspace cache tests by behavior
